Fix `ResolverTopicData` issues
